+++ common/src/main/java/earth/terrarium/argonauts/common/handlers/guild/GuildHandler.java
@@ -11,7 +11,6 @@ import earth.terrarium.argonauts.common.handlers.base.MemberException;
 import earth.terrarium.argonauts.common.handlers.base.members.Member;
 import earth.terrarium.argonauts.common.handlers.guild.members.GuildMembers;
 import earth.terrarium.argonauts.common.handlers.guild.settings.GuildSettings;
-import earth.terrarium.argonauts.common.network.NetworkHandler;
 import earth.terrarium.argonauts.common.network.messages.ClientboundSyncGuildsPacket;
 import earth.terrarium.argonauts.common.utils.EventUtils;
 import earth.terrarium.argonauts.common.utils.ModUtils;
@@ -87,7 +86,7 @@ public class GuildHandler extends SaveHandler implements GuildApi {
         guild.settings().setDisplayName(displayName);
         data.guilds.put(id, guild);
         data.playerGuilds.put(player.getUUID(), id);
-        NetworkHandler.CHANNEL.sendToAllPlayers(new ClientboundSyncGuildsPacket(Set.of(guild), Set.of()), player.server);
+        ModUtils.sendToAllClientPlayers(new ClientboundSyncGuildsPacket(Set.of(guild), Set.of()), player.server);
         player.displayClientMessage(CommonUtils.serverTranslatable("text.argonauts.member.guild_create", guild.settings().displayName().getString()), false);
 
         EventUtils.created(guild, player);
@@ -146,7 +145,7 @@ public class GuildHandler extends SaveHandler implements GuildApi {
 
         guild.members().add(player.getGameProfile());
         data.playerGuilds.put(player.getUUID(), guild.id());
-        NetworkHandler.CHANNEL.sendToAllPlayers(new ClientboundSyncGuildsPacket(Set.of(guild), Set.of()), player.server);
+        ModUtils.sendToAllClientPlayers(new ClientboundSyncGuildsPacket(Set.of(guild), Set.of()), player.server);
         player.displayClientMessage(CommonUtils.serverTranslatable("text.argonauts.member.guild_join", guild.settings().displayName().getString()), false);
 
         EventUtils.joined(guild, player);
@@ -179,7 +178,7 @@ public class GuildHandler extends SaveHandler implements GuildApi {
             serverPlayer.displayClientMessage(CommonUtils.serverTranslatable("text.argonauts.member.guild_perspective_leave", player.getName().getString(), guild.settings().displayName().getString()), false);
         }
         data.playerGuilds.remove(player.getUUID());
-        NetworkHandler.CHANNEL.sendToAllPlayers(new ClientboundSyncGuildsPacket(Set.of(guild), Set.of()), player.server);
+        ModUtils.sendToAllClientPlayers(new ClientboundSyncGuildsPacket(Set.of(guild), Set.of()), player.server);
 
         EventUtils.left(guild, player);
         if (Argonauts.isCadmusLoaded()) {
@@ -203,7 +202,7 @@ public class GuildHandler extends SaveHandler implements GuildApi {
     public void remove(boolean force, Guild guild, MinecraftServer server) {
         var data = read(server);
         data.guilds.remove(guild.id());
-        NetworkHandler.CHANNEL.sendToAllPlayers(new ClientboundSyncGuildsPacket(Set.of(), Set.of(guild.id())), server);
+        ModUtils.sendToAllClientPlayers(new ClientboundSyncGuildsPacket(Set.of(), Set.of(guild.id())), server);
         EventUtils.removed(force, guild);
         if (Argonauts.isCadmusLoaded()) {
             CadmusIntegration.disbandCadmusTeam(guild, server);